+++ addon/components/paper-select.js
@@ -4,6 +4,5 @@ import PaperInput from './paper-input';
 let { computed } = Ember;
 
 export default PaperInput.extend({
-  label: computed.alias('placeholder'),
-  onOpen: computed.alias('on-open')
+  label: computed.alias('placeholder')
 });