+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   - paper-switch now uses `value` instead of `checked`.
   - paper-select now uses required an `onchange` action.
   - paper-select now uses attribute `itemLabelCallback` instead of `item-label-callback`.
+  - paper-select now uses attribute `getItems` instead of `on-open`.
 
 ### 0.2.11